backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
chromadb
pydantic-settings
uuid
langchain
langchain-openai
langchain-community
beautifulsoup4
requests
langchain-chroma
sqlalchemy
aiomysql
redis
alembic
cryptography
pypdf
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
python-json-logger
stripe
pytest
